backend/app/requirements.txt
flask
flask-sqlalchemy
pymysql
flask-cors
